pyproject.toml
[project]
name = "Back-end"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "marcoperuzzocmd", email = "marcoantperuzzo@gmail.com"},
]
dependencies = ["django>=5.1.3", "drf-spectacular>=0.28.0", "django-restframework>=0.0.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"